Move functionality of Consultation to AdviceAnimalMacro
- Use only Advice object as prop for AdviceAnimalMacro
